// psa.service.modysservice/src/models/personalData.ts
export interface PersonalDataReq {
  anrede?: string;
  titel?: string;
  name?: string;
  vorname?: string;
  strasse?: string;
  haus_nr?: string;
  plz?: string;
  landkreis?: string;
  ort?: string;
  telefon_privat?: string;
  telefon_dienst?: string;
  telefon_mobil?: string;
  email?: string;
  comment?: string;
}